requirements.txt
requests>=2.28.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
python-dotenv>=0.19.2
